# C domain, fix performance regression since 3.0.0