--- a/sphinx/domains/c/_ast.py
+++ b/sphinx/domains/c/_ast.py
@@ -58,6 +58,10 @@ class ASTIdentifier(ASTBaseBase):
     def __str__(self) -> str:
         return self.identifier
 
+    @property
+    def name(self) -> str:
+        return self.identifier
+
     def get_display_string(self) -> str:
         return "[anonymous]" if self.is_anon() else self.identifier