@picnixz 2024-03-22
I'll wait for Jakob's review because I am not that well-versed in the C/C++ domain.