@donaldh 2024-03-22
> It's great to see a PR addressing this. If you give me a week I should be able to review it as well. I'm a bit concerned with the simplified lookup logic as there are quite some corner cases.

@jakobandersen As I understand it, the code came from the C++ domain where lookup complexity is higher due to namespaces and classes. I have tried my best to verify the lookup logic required for C. I verified the before and after generated output is the same for `make htmldocs` in the Linux kernel. If there are corner cases I have missed then we need to add them to the tests in sphinx.