@jakobandersen 2024-03-24
By the way, I fixed so the symbol debugging now works again, so a rebase onto master might be nice.