@donaldh 2024-03-25
> Unfortunately my gut feeling was right, there is some handling of duplicate symbols which changes in this PR. Test case:
> [...] 
> When a duplicate symbol is detected, then the duplicate should not overwrite the original in the dictionaries. I think it is safe to throw the duplicate symbol away, as long as the old behaviour of setting the current scope to be the original symbol is kept.

I think this is caused by `candSymbol` in `_add_symbols()`. A new symbol with a parent gets automatically added to the parent's `_childrenBy*` so this happens for `candSymbol` before the duplicate checking gets run. I did look at removing the `_add_child()` step from `__init__` and doing it after duplicate checking, but this breaks other code paths. Instead, I think it is sufficient to check for a name collision in `_add_symbols()` and never overwrite an existing entry:

```
diff --git a/sphinx/domains/c/_symbol.py b/sphinx/domains/c/_symbol.py
index 0c8905f1d59f..5194427b5466 100644
--- a/sphinx/domains/c/_symbol.py
+++ b/sphinx/domains/c/_symbol.py
@@ -123,6 +123,9 @@ class Symbol:
 
     def _add_child(self, child: Symbol) -> None:
         name = child.ident.name
+        if name in self._childrenByName:
+            # Duplicate so don't add - will be reported in _add_symbols()
+            return
         self._childrenByName[name] = child
         if child.docname not in self._childrenByDocname:
             self._childrenByDocname[child.docname] = {}
```