@donaldh 2024-03-26
While investigating another issue, I have found something that the C domain does not account for. I was looking for the root cause of these false duplicate warnings we get from Sphinx 3.0.0 thru 7.6.2:
```
/home/donaldh/net-next/Documentation/driver-api/usb/usb:164: /home/donaldh/net-next/drivers/usb/core/message.c:968: WARNING: Duplicate C declaration, also defined at driver-api/usb/gadget:802.
Declaration is '.. c:function:: int usb_string (struct usb_device *dev, int index, char *buf, size_t size)'.
/home/donaldh/net-next/Documentation/driver-api/usb/usb.rst:968: WARNING: Duplicate C declaration, also defined at driver-api/usb/gadget:802.
Declaration is '.. c:struct:: usb_string'.
/home/donaldh/net-next/Documentation/driver-api/miscellaneous:47: /home/donaldh/net-next/drivers/pwm/core.c:534: WARNING: Duplicate C declaration, also defined at driver-api/miscellaneous:224.
Declaration is '.. c:function:: int pwm_capture (struct pwm_device *pwm, struct pwm_capture *result, unsigned long timeout)'.
/home/donaldh/net-next/Documentation/gpu/drm-kms:360: /home/donaldh/net-next/drivers/gpu/drm/drm_fourcc.c:344: WARNING: Duplicate C declaration, also defined at gpu/drm-kms:39.
Declaration is '.. c:function:: const struct drm_format_info * drm_format_info (u32 format)'.
/home/donaldh/net-next/Documentation/gpu/drm-kms:461: /home/donaldh/net-next/drivers/gpu/drm/drm_modeset_lock.c:392: WARNING: Duplicate C declaration, also defined at gpu/drm-kms:49.
Declaration is '.. c:function:: int drm_modeset_lock (struct drm_modeset_lock *lock, struct drm_modeset_acquire_ctx *ctx)'.
/home/donaldh/net-next/Documentation/gpu/drm-uapi:434: /home/donaldh/net-next/drivers/gpu/drm/drm_ioctl.c:874: WARNING: Duplicate C declaration, also defined at gpu/drm-uapi:70.
Declaration is '.. c:function:: bool drm_ioctl_flags (unsigned int nr, unsigned int *flags)'.
/home/donaldh/net-next/Documentation/core-api/workqueue:776: /home/donaldh/net-next/include/linux/workqueue.h:493: WARNING: Inline literal start-string without end-string.
```
The root cause is e.g. `struct usb_string` and `int usb_string()` both being defined, which the C domain doesn't account for. Sphinx is treating this as a duplicate when it shouldn't. They should be kept distinct, within separate search spaces. One way to resolve this would be to introduce two separate search spaces:
1. Tag names – structs, unions, enums.
2. Identifier names – functions, vars, typedefs.
Question: Should I tackle that in this PR or in a separate PR?