@jakobandersen 2024-03-26
> 1. Tag names – structs, unions, enums.
> 2. Identifier names – functions, vars, typedefs.
> 
> Question: Should I tackle that in this PR or in a separate PR? The reason I ask is because it will affect the implementation used in this PR.

There is an old issue for this (https://github.com/sphinx-doc/sphinx/issues/7819), and I have a branch that should fix it (https://github.com/jakobandersen/sphinx/tree/c_ids_tags). As soon as I get time I will rebase it and get it into shape. It has been a while, but if I remember correctly it was very difficult to not break intersphinx support without https://github.com/sphinx-doc/sphinx/pull/8929. I'm currently updating it, so there can be progress.