@jakobandersen 2024-03-26
> I did look at removing the `_add_child()` step from `__init__` and doing it after duplicate checking, but this breaks other code paths. Instead, I think it is sufficient to check for a name collision in `_add_child()` and never overwrite an existing entry

I tried the same quick fix, and indeed it broke some invariants elsewhere. I think your solution for now it's a good way to fix it.