@picnixz 2024-04-13
> Looks good to me. Apparently I can't merge it without either squashing or rebasing, is that intentional?

I think our policy is to only to squash commits + write description if needed instead of rebasing + possible multiple commits.