@picnixz 2024-04-13
> Must be a "new" policy (in the last couple of years :-)).

For me it's the policy since I only arrived last year haha! But I can understand that you would like not to destroy the commits. But I think it should be fine to rebase? (and keep the commits?) 

@AA-Turner Any reason why we don't allow just using a commit message? (or can we use rebase?)